GameOver Implemented
Added the GameOverState class.

Modified the Player class to update the Player health when the kill() function is called and found a 'possible' bug that prevents the Player from losing lives at all?

Modified the update() function in PlayState to fade the camera to black and call a helper function that changes the current game state to GameOverState.